Add functionality to account for spaces in cityName. Add CSS

+++ WeatherAPI/src/main/java/com/raphael/WeatherAPI/model/CurrentWeather.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 public record CurrentWeather(String location,
-                             @JsonProperty("temp") Double temperature,
+                             @JsonProperty("temp") int temperature,
                              @JsonProperty("humidity") int humidity,
                              @JsonProperty("speed") Double windSpeed,
                              @JsonProperty("description") String description) {
